вносит правки
