+++ src/presenter/trip-events-filter-presenter.js
@@ -1,10 +1,10 @@
 import {render, RenderPosition} from '../render.js';
 import TripEventsFormView from '../view/trip-events-sort-form-view.js';
-import TripSortDayView from '../view/trip-sort-day-view.js';
-import TripSortEventView from '../view/trip-sort-event-view.js';
-import TripSortTimeView from '../view/trip-sort-time-view.js';
-import TripSortPriceView from '../view/trip-sort-price-view.js';
-import TripSortOfferView from '../view/trip-sort-offer-view.js';
+import TripSortView from '../view/trip-sort-view.js';
+// import TripSortEventView from '../view/trip-sort-event-view.js';
+// import TripSortTimeView from '../view/trip-sort-time-view.js';
+// import TripSortPriceView from '../view/trip-sort-price-view.js';
+// import TripSortOfferView from '../view/trip-sort-offer-view.js';
 
 const tripEventsSectionElement = document.querySelector('.trip-events');
 
@@ -17,11 +17,11 @@ export default class TripEventsFilterPresenter {
 
   init() {
     render(this.#tripEventsFormElement, tripEventsSectionElement, RenderPosition.AFTERBEGIN);
-    render(new TripSortDayView(), this.#tripEventsFormElement.element);
-    render(new TripSortEventView(), this.#tripEventsFormElement.element);
-    render(new TripSortTimeView(), this.#tripEventsFormElement.element);
-    render(new TripSortPriceView(), this.#tripEventsFormElement.element);
-    render(new TripSortOfferView(), this.#tripEventsFormElement.element);
+    render(new TripSortView(), this.#tripEventsFormElement.element);
+    // render(new TripSortEventView(), this.#tripEventsFormElement.element);
+    // render(new TripSortTimeView(), this.#tripEventsFormElement.element);
+    // render(new TripSortPriceView(), this.#tripEventsFormElement.element);
+    // render(new TripSortOfferView(), this.#tripEventsFormElement.element);
   }
 
   #handleSortTypeChange(sortType) {
